refactor scoreboard store into a module
